pc-client/requirements.txt
# MobileAgentLivelink PC Client
# Python 3.10+

websockets>=12.0
pyautogui>=0.9.54
pygetwindow>=0.0.9
pynput>=1.7.6
python-dotenv>=1.0.0
